% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mail_recipients.R
\name{add_mail_recipients}
\alias{add_mail_recipients}
\title{Getting the project module IDs and their titles}
\usage{
add_mail_recipients(
  participant_events,
  mail_recipient_codes = tibble(recipient = character(), code = character())
)
}
\arguments{
\item{participant_events}{Tibble including minimally pre-processed log data for a single participation}

\item{mail_recipient_codes}{Tibble including previously assigned mail recipients and their codes}
}
\value{
A tibble including a completed list of mail recipients used by the participants and their codes for generating
more generalized event codes
}
\description{
Takes the event data from a single participant and returns a table with mail recipients the participant was writing a mail and 4-digit codes
for each recipient
}
