% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_user_created_mails.R
\name{add_user_created_mails}
\alias{add_user_created_mails}
\title{Adding User Created Mails to the Scenario Elements}
\usage{
add_user_created_mails(scenario_elements, participant_events)
}
\arguments{
\item{scenario_elements}{Table including currently existing scenario elements and their codes}

\item{participant_events}{Table including minimally pre-processed log data for a single participation}
}
\value{
Table including a completed list of scenario elements
}
\description{
Takes the event data from a single participant and returns a table with mail
recipients the participant was writing a mail and 4-digit codes for each
recipient.
}
