% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_list.R
\name{aggregate_duplicates}
\alias{aggregate_duplicates}
\title{Aggregate event data from a single participation}
\usage{
aggregate_duplicates(event_list)
}
\arguments{
\item{event_list}{A list including the event data}
}
\value{
A list including the aggregated event data
}
\description{
Takes event data from a single participation and returns an aggregated form,
where events with identical codes that occur directly after each other are
aggregated to a single event. The duration values are correspondingly adjusted,
and for each aggregated event an intensity is calculated that describes how
many events occurred in the original form.
}
