% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_events.R
\name{aggregate_events}
\alias{aggregate_events}
\title{Aggregate Event Data}
\usage{
aggregate_events(survey_data, substr_start = 1, substr_stop = 10, interval = 0)
}
\arguments{
\item{survey_data}{LUCA survey data object}

\item{substr_start}{First character element to be considered in the name
of the event}

\item{substr_stop}{Last character element to be considered in the name
of the event}

\item{interval}{Number of seconds for each considered interval}
}
\value{
A list including the aggregated event data
}
\description{
Takes the event data from a LUCA survey data object and aggregates them,
either according to parts of the name, by a time interval, or by
both.
}
\details{
In the first case \code{substr_start} marks the starting element of the substring
of the event name that is considered as relevant and \code{substr_stop} marks the
corresponding last character.
In the second case \code{interval} defines the number of seconds that will be
used to split the maximum module time into intervals; each interval will
then be assigned the event code that lasted the longest within the given
interval. As a result the event lists of all participants will be of equal
length.
Both variants can also be combined.
}
