% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_events.R
\name{aggregate_events_internal}
\alias{aggregate_events_internal}
\title{Aggregate event data from a single participation}
\usage{
aggregate_events_internal(..., substr_start, substr_stop, interval)
}
\arguments{
\item{...}{List of arguments (in fact columns) of the event data from a single participation}

\item{substr_start}{First character element to be considered in the name
of the event}

\item{substr_stop}{Last character element to be considered in the name
of the event}

\item{interval}{Number of seconds for each considered interval}
}
\value{
A list including the aggregated event data
}
\description{
Takes event data from a single participation (event codes, times, and durations)
and returns an aggregatd form of the three vectors within a table.
}
