% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basic_wf_codes}
\alias{basic_wf_codes}
\title{Basic workflow codes.}
\format{
A tibble with XX rows and YY variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
Constructed manually based on the unique event types provided in the log data of LUCA Office.
}
\usage{
basic_wf_codes
}
\description{
A tibble providing the basic workflow codes and descriptions for the different event types in the LUCA Office log data.
}
\keyword{datasets}
