% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_events.R
\name{compress_events}
\alias{compress_events}
\title{Compresses the event data from a single participation}
\usage{
compress_events(workflow)
}
\arguments{
\item{workflow}{A list including the event data}
}
\value{
A list including the compressed event data
}
\description{
Takes the event data from a single participation and returns a compressed form,
where events with identical codes that occur directly after each other are
compressed to a single event. The durations are correspondingly adjusted,
and for each compressed event an intensity is calculated that describes how
many events occurred in the original form.
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
workflow <- get_workflow(json_data)
compressed_workflow <- compress_events(workflow)
}

}
