% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenario_elements.R
\name{construct_element_code}
\alias{construct_element_code}
\title{Helper function to assign the running event code for elements of the same type}
\usage{
construct_element_code(relevance)
}
\arguments{
\item{relevance}{string indicating the relevance of a given LUCA office element}
}
\description{
Helper function to assign the running event code for elements of the same type
}
