% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_elements.R
\name{construct_wf_code}
\alias{construct_wf_code}
\title{Helper function to assign the running workflow code for elements of the same type}
\usage{
construct_wf_code(relevance)
}
\arguments{
\item{relevance}{string indicating the relevance of a given LUCA office element}
}
\description{
Helper function to assign the running workflow code for elements of the same type
}
