% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{document_type_coding}
\alias{document_type_coding}
\title{Document type coding.}
\format{
A tibble with XX rows and YY variables:
\describe{
\item{document_id}{project specific document ID, a 3 digit number}
\item{code}{a two digit document type code that is assigned to the given document_id}
\item{label}{a label for the document type}
}
}
\source{
Constructed manually based on the unique event types provided in the log data of LUCA Office.
}
\usage{
document_type_coding
}
\description{
A dataframe providing a two digit code to additionally assign each document
to a specific document type
}
\keyword{datasets}
