% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_events.R
\name{get_intensity_sum}
\alias{get_intensity_sum}
\title{Aggregate intensity values in a new vector}
\usage{
get_intensity_sum(intensity)
}
\arguments{
\item{intensity}{Vector including intensity values}
}
\description{
The aggregated vector includes at each position the sum of all intensity
values provided until that positions in the given intensity vector
}
