% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participation_data.R
\name{get_participation_data}
\alias{get_participation_data}
\title{Getting the General Participation Information}
\usage{
get_participation_data(json_data)
}
\arguments{
\item{jsondata}{The log data for a single participation in form of a json object}
}
\value{
A dataframe (consisting one row) including general information for a single participation
}
\description{
Takes the log data from a single participation and returns a dataframe including a single line with general information of the participation.
}
\examples{

}
