% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_elements.R
\name{get_project_elements}
\alias{get_project_elements}
\title{Getting the scenario elements and its respective codes for the data}
\usage{
get_project_elements(json_data, basic_wf_codes = basic_wf_codes)
}
\arguments{
\item{json_data}{The log data for a single participation in form of a nested list}
}
\value{
A dataframe including the scenario elements, their workflow codes and other relevant information
}
\description{
Takes the log data from a single participation and returns a list with the names of all project elements and the respective workflow codes.
These might prepared onboarding steps, emails, excel sheets, pdf, events and other elements.
}
\examples{

}
