% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_modules.R
\name{get_project_modules}
\alias{get_project_modules}
\title{Getting the project module IDs and their titles}
\usage{
get_project_modules(json_data)
}
\arguments{
\item{json_data}{Nested list including the log data for a single participation}
}
\value{
A dataframe including information on the project modules: module code, module title, hash ID of the module, and hash ID of the sample company used in the modules
}
\description{
Takes the log data from a single participation and returns a table with the IDs and names of all modules in the project
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
project_modules <- get_project_modules(json_data)
}

}
