% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_scenarios.R
\name{get_project_scenarios}
\alias{get_project_scenarios}
\title{Getting the project scenario IDs and their titles}
\usage{
get_project_scenarios(json_data, hash_ids = FALSE)
}
\arguments{
\item{json_data}{Nested list including the log data for a single participation}

\item{hash_ids}{If TRUE, the returned dataframe includes the hash IDs of the scenarios}
}
\value{
A dataframe including information on the project scenarios: scenario workflow code, scenario title, hash ID of the scenario, and hash ID of the sample company used in the scenario
}
\description{
Takes the log data from a single participation and returns a table with the IDs and names of all scenarios in the project
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
project_scenarios <- get_project_scenarios(json_data)
}

}
