% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questionnaire_data.R
\name{get_questionnaire_data}
\alias{get_questionnaire_data}
\title{Answer Data of the Questionnaires from a Single Participation}
\usage{
get_questionnaire_data(
  json_data,
  project_modules,
  scenario_elements,
  questionnaire_elements,
  rater,
  debug_mode = FALSE,
  full_rater_data = FALSE
)
}
\arguments{
\item{json_data}{The log data for a single participation in form of a json object}

\item{project_modules}{Dataframe including the running numbers and ids for the project modules}

\item{scenario_elements}{Dataframe including the codes for all existing scenario elements, which are used in the event list.}

\item{questionnaire_elements}{Dataframe including the codes for all questionnaire elements, which are used in the event list as well as in the summary results.}

\item{rater}{Dataframe including the running numbers and ids for the rater in the project}

\item{debug_mode}{If TRUE the internal hash IDs for the project elements are included and no module specific data is returned.}

\item{full_rater_data}{If FALSE only the final score will be included for the questionnaire data. If TRUE all scores from all raters will be included.}
}
\value{
A dataframe (consisting of one row) including general information for a single participation
}
\description{
Takes the log data and the already prepared workflow data from a single participant
and returns a single line dataframe with the answer data of included questionnaires.
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
workflow <- get_workflow(json_data, module_specific=TRUE)
questionnaire_data <- get_questionnaire_data(json_data, workflow)
}

}
