% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questionnaire_elements.R
\name{get_questionnaire_elements}
\alias{get_questionnaire_elements}
\title{Getting the questionnaire elements from a project}
\usage{
get_questionnaire_elements(json_data, hash_ids = FALSE)
}
\arguments{
\item{json_data}{Nested list including the log data for a single participation}

\item{hash_ids}{If TRUE the internal hash IDs for the questionnaire elements are included}
}
\value{
A dataframe including all questionnaire elements, their event codes and other relevant information
}
\description{
Takes the log data from a single participation and returns a table with the
names of all questionnaire elements and runtime survey elements (which are
technically questionnaires as well) and the respective event codes.
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
questionnaire_elements <- get_questionnaire_elements(json_data)
}

}
