% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rater.R
\name{get_rater}
\alias{get_rater}
\title{Getting the list of raters assigned in this project}
\usage{
get_rater(json_data, hash_ids = FALSE)
}
\arguments{
\item{json_data}{Nested list including the log data for a single participation}

\item{hash_ids}{If TRUE the internal hash IDs for the questionnaire elements are included}
}
\value{
A dataframe including all rater, their id number, names, emails, and organizations
}
\description{
Takes the log data from a single participation and returns a table with the names of all raters and their assigned rater number.
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
rater <- get_rater(json_data)
}

}
