% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_companies.R
\name{get_sample_companies}
\alias{get_sample_companies}
\title{Getting the sample companies, their titles, and description}
\usage{
get_sample_companies(json_data)
}
\arguments{
\item{json_data}{Nested list including the log data for a single participation}
}
\value{
A dataframe including information on the sample companies: their respective (hash) \code{id}, \code{title}, and \code{description}
}
\description{
Takes the log data from a single participation and returns a table with the sample company information
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
project_modules <- get_sample_companies(json_data)
}

}
