% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workflow.R
\name{get_workflow}
\alias{get_workflow}
\title{Getting the workflow data from a single participation}
\usage{
get_workflow(json_data)
}
\arguments{
\item{jsondata}{The log data for a single participation in form of a nested list}
}
\value{
A list including the workflow data
}
\description{
Takes the log data from a single participation and returns a list describing the workflow.
}
\examples{

}
