% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{import_basic_coding}
\alias{import_basic_coding}
\title{Helper function to import tables including the basic coding information and
add them to the package project.}
\usage{
import_basic_coding(file = "dev/basic_codes.xlsx")
}
\arguments{
\item{file}{Name of the Excel file including the tables with the standard basic workflow codes, the standard tool codes and example table for defining project specific document types (e.g. considering their general content)}
}
\description{
The function is only used during development of the package.
}
