% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_idle_events.R
\name{insert_idle_events}
\alias{insert_idle_events}
\title{Inserts event that are marked as idle}
\usage{
insert_idle_events(workflow, idle_time = 5)
}
\arguments{
\item{workflow}{A list including the event data}

\item{idle_time}{Numeric describing after how many seconds an event is considered as idle.}
}
\value{
An event list including additional idle events
}
\description{
Takes the event data from a single participation and returns the data including
additional events that are marked as idle when the participant was not active
for more than \code{idle_time} seconds.
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
workflow <- get_workflow(json_data)
workflow_with_idle_events <- insert_idle_events(workflow)
}

}
