% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucar-package.R
\docType{package}
\name{lucar-package}
\alias{lucar}
\alias{lucar-package}
\title{lucar: Prepare Survey Data from LUCA Office}
\description{
Functions to prepare survey data exported from the LUCA Office application (https://luca-office.de/) for secondary analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LucaOffice/lucar}
}

}
\author{
\strong{Maintainer}: Steffen Brandt \email{steffen@opencampus.sh} (\href{https://orcid.org/0000-0001-5410-7186}{ORCID})

}
\keyword{internal}
