% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_lsd.R
\name{prepare_lsd}
\alias{prepare_lsd}
\title{Prepares Survey Data from LUCA Office for secondary Analyses}
\usage{
prepare_lsd(
  path = "./",
  aggregate_duplicate_events = FALSE,
  idle_time = 10,
  unzip = TRUE,
  event_codes = lucar::event_codes,
  tool_codes = lucar::tool_codes,
  debug_mode = FALSE,
  full_rater_data = FALSE
)
}
\arguments{
\item{path}{Path to the folder including zip archives downloaded from LUCA
Office (already unzipped JSON files as well as archives or files in
subfolders are also considered).}

\item{aggregate_duplicate_events}{For \code{TRUE}, events with identical codes
directly following each other will be collated to a single event.}

\item{idle_time}{For values larger than 0, events will be marked as idle,
if the event following this one happened more than \code{ìdle_time} seconds
later. This might be relevant for analyses considering times when
participants were actually not doing anything.}

\item{unzip}{For \code{TRUE}, zip archives located in the given path and
following the naming structure from LUCA Office will be unzipped; for
\code{FALSE}, only JSON files from previously unzipped archives will be
considered.}

\item{event_codes}{Table with event codes that are used to identify and
differentiate events. The package provides a standard encoding of the
events, however, for very specific analyses it can be adapted as needed.}

\item{tool_codes}{Table with codes for the different available tools. These
are used similar as the event codes to identify and differentiate events.
The package also provides a standard encoding of the events. For very
specific analyses it can also be adapted as needed.}

\item{debug_mode}{For \code{TRUE} the returned tables include internal hash IDs
for the different project elements (to check on possible matching errors).
Further, the event list is not separated according to the different
modules but only a single event list for the complete project, and an
additional table including possibly unrecognized events (missing in the
\code{event_codes} or \code{tool_codes}) is returned.}

\item{full_rater_data}{If FALSE only the final score will be included for
the questionnaire data. If TRUE all scores from all raters will be included.}
}
\value{
List including different tables (tibbles) with participation data
and documentation on the administered project.
}
\description{
The survey data collected via LUCA Office can be downloaded in form of a zip
archive that includes JSON data for each participation. \code{prepare_lsd()}
prepares the nested JSON data from the participations in a summarized form
for secondary analyses.
}
\details{
Besides a table including the relevant information for each participation,
the function additionally returns different tables with overviews on the
elements used in the scenarios, the form of included questionnaires, and
the raters responsible for the scoring.
}
\examples{

\dontrun{
# Prepare survey data given in the current directory
survey_data <- prepare_lsd()
}

}
