% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read_json}
\alias{read_json}
\title{Reads all JSON files into a single dataframe}
\usage{
read_json(path, unzip = FALSE)
}
\arguments{
\item{path}{The path of the folder including all JSON files (might be included in subfolders).}
}
\value{
A dataframe including the prepared data from all JSON files
}
\description{
JSON files exported from the LUCA office simulation into a dedicated folder
are read into a datafrmame.
}
