% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tool_coding}
\alias{tool_coding}
\title{Standard tool coding.}
\format{
A tibble with XX rows and YY variables:
\describe{
\item{tool}{name of the tool in the LUCA office log data}
\item{code}{a two digit code that is assigned the given tool}
\item{label}{a label for the tool}
}
}
\source{
Constructed manually based on the unique event types provided in the log data of LUCA Office.
}
\usage{
tool_coding
}
\description{
A dataframe providing two digit codes for each tool available in LUCA office
}
\keyword{datasets}
