% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{workflow_coding}
\alias{workflow_coding}
\title{Standard workflow coding}
\format{
A tibble with XX rows and YY variables:
\describe{
\item{event_type}{all of the currently 44 different log event types}
\item{label}{label for the event type}
\item{wf_code}{a code of 9 uppercase characters characterizing a given log event}
\item{data_example}{supplement data for each log event, in json format}
\item{index}{running number of the event as given in the original log table}
}
}
\source{
Constructed manually based on the unique event types provided in the log data of LUCA Office.
}
\usage{
workflow_coding
}
\description{
A tibble providing the basic workflow codes and descriptions for the different event types in the LUCA Office log data.
}
\details{
Events with wf_codes defined as '#' are ignored from the preparation of the workflow data
}
\keyword{datasets}
