% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_logdata.R
\name{write_logdata_csv}
\alias{write_logdata_csv}
\alias{write_logdata_csv2}
\title{Write LUCA Office logdata data to csv files}
\usage{
write_logdata_csv(logdata, folder = "logdata")

write_logdata_csv2(logdata, folder = "logdata")
}
\arguments{
\item{logdata}{A list including prepared data from the function \code{prepare_logdata()}.}

\item{folder}{A folder to which the csv files including the workflow data are exported.}
}
\value{
A list including the prepared data from all JSON files
}
\description{
After preparing (JSON) log data from LUCA office using  \code{prepare_data()}, you can
use \code{write_logdata_csv()} or \code{write_logdata_csv2()} to write the workflow data, the
participation data including aggregated information on the person level as well as
two tables including information on the files existing in the modules and the
names of the modules themselves.
}
\examples{

# Searches in the current working directory and all subdirectories for log data from LUCA office
# and prepares the data, which is then use in write_workflow_data to save the part including the
# workflow data as csv
\dontrun{
logdata <- prepare_logdata()
write_workflow_csv(logdata)
}

}
