% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lsd.R
\name{write_lsd}
\alias{write_lsd}
\title{Save Survey Data from LUCA Office Using a Specified Write Function}
\usage{
write_lsd(survey_data, export_folder = "lsd_csv", write_function)
}
\arguments{
\item{survey_data}{List including the result object from the function
\code{prepare_lsd()}.}

\item{export_folder}{Folder to which the csv files including the workflow data
are exported.}

\item{write_function}{Function which is used to save the different tables
included in the prepared survey data.}
}
\value{
\code{write_lsd()} returns the input `survey_data' invisibly.
}
\description{
This function does the heavy lifting for the write functions
\code{write_lsd_csv()} and \code{write_lsd_csv2()} and allows to flexibly define
additional functions to save the survey data.
}
