% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lsd.R
\name{write_lsd_csv}
\alias{write_lsd_csv}
\alias{write_lsd_csv2}
\title{Save Survey Data from LUCA Office in CSV Files}
\usage{
write_lsd_csv(survey_data, export_folder = "lsd_csv")

write_lsd_csv2(survey_data, export_folder = "lsd_csv")
}
\arguments{
\item{survey_data}{List including the result object from the function
\code{prepare_lsd()}.}

\item{export_folder}{Folder to which the csv files including the workflow data
are exported.}
}
\value{
\verb{write_lsd_*()} returns the input `survey_data' invisibly.
}
\description{
After preparing survey data from LUCA office using \code{prepare_lsd()}, you
can use \code{write_lsd_csv()} or \code{write_lsd_csv2()} to save the various tables
returned from \code{prepare_lsd()} in form of CSVs,
}
\examples{

# Searches in the current working directory and all subdirectories for log data from LUCA office
# and prepares the data, which is then use in write_workflow_data to save the part including the
# workflow data as csv
\dontrun{
survey_data <- prepare_lsd()
write_lsd_csv2(survey_data)
}

}
