% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{write_unique_events}
\alias{write_unique_events}
\title{Helper function to construct a table including the information for the workflow coding}
\usage{
write_unique_events(events, file = "dev/unique_events.csv")
}
\arguments{
\item{events}{Dataframe including log events from LUCA office}

\item{file}{Name of the csv file for saving the results}
}
\description{
It provides all unique log events provided in the events table and
formats the output table in the form needed to define the basic
workflow table (see also function import_basic_coding).
}
\details{
The function is only used during development of the package.
}
