% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_events.R
\name{aggregate_events}
\alias{aggregate_events}
\title{Aggregates the event data from a single participation}
\usage{
aggregate_events(workflow)
}
\arguments{
\item{workflow}{A list including the event data}
}
\value{
A list including the compressed event data
}
\description{
Takes the event data from a single participation and returns an aggregated form,
where events with identical codes that occur directly after each other are
aggregated to a single event. The duration values are correspondingly adjusted,
and for each aggregated event an intensity is calculated that describes how
many events occurred in the original form.
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
workflow <- get_workflow(json_data)
compressed_workflow <- aggregate_events(workflow)
}

}
