% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participation_data.R
\name{get_participation_data}
\alias{get_participation_data}
\title{Getting the General Participation Information}
\usage{
get_participation_data(json_data, hash_ids = FALSE)
}
\arguments{
\item{json_data}{The log data for a single participation in form of a json object}

\item{hash_ids}{If TRUE the internal hash IDs for the projct elements are included}
}
\value{
A dataframe (consisting one row) including general information for a single participation
}
\description{
Takes the log data from a single participation and returns a dataframe including a single line with general information of the participation.
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
participation_data <- get_participation_data(json_data)
}

}
