% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_elements.R
\name{get_project_elements}
\alias{get_project_elements}
\title{Getting the scenario elements and its respective codes for the data}
\usage{
get_project_elements(json_data, hash_ids = FALSE)
}
\arguments{
\item{json_data}{Nested list including the log data for a single participation}

\item{hash_ids}{If TRUE the internal hash IDs for the projct elements are included}
}
\value{
A dataframe including the scenario elements, their workflow codes and other relevant information
}
\description{
Takes the log data from a single participation and returns a list with the names of all project elements and the respective workflow codes.
These might prepared onboarding steps, emails, excel sheets, pdf, events and other elements.
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
project_elements <- get_project_elements(json_data)
}

}
