% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logdata_summary.R, R/prepare_lsd.R
\name{getTime}
\alias{getTime}
\title{Formatting time data}
\usage{
getTime(time, tzone = "CET")

getTime(time, tzone = "CET")
}
\arguments{
\item{time}{string including the time information}

\item{tzone}{string defining the time zone that will be use to format the time}
}
\value{
Timing data including the provided time zone definition
}
\description{
Formatting time data

Format timing data
}
