\name{lucar-package}
\alias{lucar-package}
\alias{lucar}
\docType{package}
\title{
\packageTitle{lucar}
}
\description{
\packageDescription{lucar}
}
\author{
\packageAuthor{lucar}

Maintainer: \packageMaintainer{lucar}
}
\references{
Literature or other references for background information
}
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
}
\examples{

#(simple examples of the most important functions)

# Searches in the current working directory and all subdirectories for log data from LUCA office
# and prepares the data in a structure suitable for further analyses
\dontrun{
log_data <- read_json()
}

}
