\name{lucar-package}
\alias{lucar-package}
\alias{lucar}
\docType{package}
\title{
\packageTitle{lucar}
}
\description{
\packageDescription{lucar}
}
\author{
\packageAuthor{lucar}

Maintainer: \packageMaintainer{lucar}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
