% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read_json}
\alias{read_json}
\title{Reads all JSON files into a single dataframe}
\usage{
read_json(path = "./", unzip = FALSE)
}
\arguments{
\item{path}{The path to the folder including all JSON files (files in subfolders are also considered).}

\item{unzip}{If true, the function looks for zip archives located in the given path, corresponding to the naming convention for exported data from LUCA Office, and unzips these.}
}
\value{
A dataframe including the prepared data from all JSON files
}
\description{
JSON files exported from the LUCA office simulation into a dedicated folder
are read into a datafrmame.
Best is to dowload all zip files into a dedicated folder and unpack them in that same folder.
Then provide this folder's path to the function.
}
\examples{

}
