% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_workflow.R
\name{summarize_workflow}
\alias{summarize_workflow}
\title{Summarizing the workflow data from a single participation}
\usage{
summarize_workflow(workflow)
}
\arguments{
\item{workflow}{A list including the workflow data}
}
\value{
A list including the summarized workflow data
}
\description{
Takes the workflow data from a single participation and returns a summarized form,
where events with identical workflow codes that occur directly after each other are
summarized to a single activity. The durations are correspondingly returned on the
activity level, and for each task an intensity is calculated that describes how
many events occurred in the original datasets within a given activity.
}
\examples{
\dontrun{
json_file = "participation_logdata.json"
json_data <- rjson::fromJSON(json_file)
workflow <- get_workflow(json_data)
sum_workflow <- summarize(workflow)
}

}
